inappropriate usize

+++ exercises/difference-of-squares/example.rs
@@ -1,12 +1,12 @@
-pub fn square_of_sum(n: usize) -> usize {
+pub fn square_of_sum(n: u32) -> u32 {
     let sum = n * (n + 1) / 2;
     sum * sum
 }
 
-pub fn sum_of_squares(n: usize) -> usize {
+pub fn sum_of_squares(n: u32) -> u32 {
     (0..n + 1).map(|x| x * x).fold(0, |accum, x| accum + x)
 }
 
-pub fn difference(n: usize) -> usize {
+pub fn difference(n: u32) -> u32 {
     square_of_sum(n) - sum_of_squares(n)
 }
