+++ exercises/dominoes/src/lib.rs
@@ -1,3 +1,3 @@
-pub fn chain(input: &[(usize, usize)]) -> Option<Vec<(usize, usize)>> {
+pub fn chain(input: &[(u8, u8)]) -> Option<Vec<(u8, u8)>> {
     unimplemented!("From the given input '{:?}' construct a proper dominoes chain or return None if it is not possible.", input);
 }
