+++ exercises/spiral-matrix/src/lib.rs
@@ -1,4 +1,4 @@
-pub fn spiral_matrix(size: usize) -> Vec<Vec<usize>> {
+pub fn spiral_matrix(size: u32) -> Vec<Vec<u32>> {
     unimplemented!(
         "Function that returns the spiral matrix of square size {}",
         size
